% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{angle_distribution}
\alias{angle_distribution}
\title{Analyze edge distribution}
\usage{
angle_distribution(im, bin.width = 3, is.orientation = TRUE)
}
\arguments{
\item{im}{an image}

\item{bin.width}{bin width in degree}

\item{is.orientation}{TRUE for orientation, FALSE for direction}
}
\description{
Analyze edge distribution
}
\examples{
df = angle_distribution(regatta)
plot(df$Angle, df$Magnitude, type = "h")
}
