% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{apply_oval_mask}
\alias{apply_oval_mask}
\title{Add an oval mask to an image}
\usage{
apply_oval_mask(im, color = 0.5)
}
\arguments{
\item{im}{an image}

\item{color}{Color of the margin area}
}
\description{
Add an oval mask to an image
}
\examples{
plot(apply_oval_mask(regatta, 0.5))
plot(apply_oval_mask(regatta, c(1,1,0.2)))
}
