% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{as.data.frame.nimg}
\alias{as.data.frame.nimg}
\title{nimg to data frame conversion}
\usage{
\method{as.data.frame}{nimg}(x, ..., is.wide = TRUE)
}
\arguments{
\item{x}{an image}

\item{...}{S3 generic/method consistency}

\item{is.wide}{either TRUE for wide format or FALSE for long format}
}
\value{
a data frame
}
\description{
nimg to data frame conversion
}
