% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{box_blur}
\alias{box_blur}
\title{Box blur}
\usage{
box_blur(im, radius)
}
\arguments{
\item{im}{an image}

\item{radius}{radius}
}
\value{
an image
}
\description{
Box blur
}
\examples{
plot(box_blur(regatta, 10))
}
