% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{box_variance}
\alias{box_variance}
\title{Box variance}
\usage{
box_variance(im, radius)
}
\arguments{
\item{im}{an image}

\item{radius}{radius}
}
\value{
an image
}
\description{
Box variance
}
\examples{
plot(box_variance(regatta, 3), rescale = TRUE)
}
