% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{CIELAB_visualize}
\alias{CIELAB_visualize}
\title{Visualize CIELAB image}
\usage{
CIELAB_visualize(im, scale = FALSE)
}
\arguments{
\item{im}{an RGB image}

\item{scale}{either TRUE or FALSE (default)}
}
\value{
a list of images
}
\description{
Visualize CIELAB image
}
