% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{circular_color_stat}
\alias{circular_color_stat}
\title{Circular color statistics}
\usage{
circular_color_stat(im)
}
\arguments{
\item{im}{an image}
}
\description{
Circular color statistics
}
\examples{
df = circular_color_stat(regatta)
}
