% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{color_dodge}
\alias{color_dodge}
\title{Color dodge}
\usage{
color_dodge(bottom, top = bottom)
}
\arguments{
\item{bottom}{an image}

\item{top}{an image}
}
\value{
an image
}
\description{
Color dodge
}
\examples{
plot(color_dodge(regatta, regatta^10))
}
