% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{colorder}
\alias{colorder}
\title{change the column order of a data frame}
\usage{
colorder(df, name)
}
\arguments{
\item{df}{a data frame}

\item{name}{column names}
}
\value{
a data frame
}
\description{
change the column order of a data frame
}
