% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{create_dead_leaves}
\alias{create_dead_leaves}
\title{Create a 'dead-leaves' image}
\usage{
create_dead_leaves(
  height,
  width = height,
  shape = "square",
  grayscale = FALSE,
  sigma_n = 2,
  rmin = 0.02,
  rmax = 0.5
)
}
\arguments{
\item{height}{image height}

\item{width}{image width}

\item{shape}{the shape of leaves. either "square" (default) or "disk"}

\item{grayscale}{either FALSE (default, color output) or TRUE (grayscale output)}

\item{sigma_n}{numeric}

\item{rmin}{minimum diameter of each leaf}

\item{rmax}{maximum diameter of each leaf}
}
\value{
an image
}
\description{
Create a 'dead-leaves' image
}
\examples{
plot(create_dead_leaves(128))
}
