% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{cubic_spline}
\alias{cubic_spline}
\title{Calculate a cubic spline}
\usage{
cubic_spline(x, low = 0, high = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{low}{minimum value of output}

\item{high}{maximum value of output}
}
\value{
a numeric vector
}
\description{
Calculate a cubic spline
}
\examples{
x = seq( from = 0, to = 1, by = 0.01 )
plot( x, cubic_spline( x ) )
}
