% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{edge_density}
\alias{edge_density}
\title{Edge density of an image}
\usage{
edge_density(im, method = "sobel")
}
\arguments{
\item{im}{an image}

\item{method}{edge detection method. either "diff", "sobel", "prewitt", or "canny".}
}
\value{
a numeric value
}
\description{
Edge density of an image
}
\examples{
edge_density(regatta)
}
