% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{edge_diff}
\alias{edge_diff}
\title{Edge detection by finite differences}
\usage{
edge_diff(im, pad.method = "mirror")
}
\arguments{
\item{im}{an image}

\item{pad.method}{either "zero", "mean", "repeat", "mirror", or a numeric value}
}
\value{
a list containing images
}
\description{
Edge detection by finite differences
}
\examples{
edge = edge_diff(im_gray(regatta))
plot(edge$magnitude)
}
