% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{edge_sobel}
\alias{edge_sobel}
\title{Sobel edge detection}
\usage{
edge_sobel(im, pad.method = "mirror")
}
\arguments{
\item{im}{an image}

\item{pad.method}{either "zero", "mean", "repeat", "mirror", or a numeric value}
}
\value{
a list containing images
}
\description{
Sobel edge detection
}
\examples{
edge = edge_sobel(im_gray(regatta))
plot(edge$magnitude)
plot(1 - edge$magnitude, rescale = TRUE)
}
