% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{edge_variance}
\alias{edge_variance}
\title{Edge detection by variance method}
\usage{
edge_variance(im, radius, pad.method = "mirror")
}
\arguments{
\item{im}{an image}

\item{radius}{kernel radius}

\item{pad.method}{either "zero", "mean", "repeat", "mirror", or a numeric value}
}
\value{
an image
}
\description{
Edge detection by variance method
}
\examples{
edge = edge_variance(im_gray(regatta), radius = 1)
plot(edge, rescale = TRUE)
}
