% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{eucdist}
\alias{eucdist}
\title{Calculate the Euclidean distance between two points}
\usage{
eucdist(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x coordinate of point 1}

\item{y1}{y coordinate of point 1}

\item{x2}{x coordinate of point 2}

\item{y2}{y coordinate of point 2}
}
\value{
distance between point 1 and 2
}
\description{
Calculate the Euclidean distance between two points
}
\examples{
eucdist(0, 0, 2, 2)
}
