% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fade_border}
\alias{fade_border}
\title{Add a fading border to an image}
\usage{
fade_border(im, npix, blur_radius = 6)
}
\arguments{
\item{im}{an image}

\item{npix}{thickness of the fading border}

\item{blur_radius}{apply a blur of this radius to the fading border to avoid illusory edges at the image corners}
}
\value{
an image
}
\description{
Add a fading border to an image
}
\examples{
plot(fade_border(regatta, npix = 30))
}
