% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fade_mask}
\alias{fade_mask}
\title{Add a fading border to an image region}
\usage{
fade_mask(im, mask, npix, blur_radius = 3)
}
\arguments{
\item{im}{an image}

\item{mask}{an image}

\item{npix}{thickness of the fading border}

\item{blur_radius}{apply a blur of this radius to the fading border to avoid illusory edges at the image corners}
}
\value{
an image
}
\description{
Add a fading border to an image region
}
\examples{
im = regatta
mask = im_mono(0, im_height(im), im_width(im))
mask[ 100:350, 100:500,] = 1
plot(fade_mask(im, mask, 20))
}
