% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fft_amplitude}
\alias{fft_amplitude}
\title{Calculate fft amplitude}
\usage{
fft_amplitude(im, modify = FALSE, dc_value)
}
\arguments{
\item{im}{an image}

\item{modify}{if TRUE, amplitude value is log-rescaled, which is useful for plotting purpose}

\item{dc_value}{if given, DC (direct current) component is replaced with this value}
}
\value{
an image
}
\description{
Calculate fft amplitude
}
\examples{
amp = fft_amplitude(im_gray(regatta), modify = TRUE)
plot(amp)
}
