% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fft_amplitude1D}
\alias{fft_amplitude1D}
\title{Calculate orientation-averaged amplitude}
\usage{
fft_amplitude1D(im, range = c(10, 512), n = 20, mask = F)
}
\arguments{
\item{im}{an image}

\item{range}{frequency range}

\item{n}{number of samples}

\item{mask}{if TRUE, apply circular mask before calculation}
}
\value{
a data frame
}
\description{
Calculate orientation-averaged amplitude
}
\examples{
df = fft_amplitude1D(regatta)
plot(df)
}
