% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fft_angular_mask}
\alias{fft_angular_mask}
\title{Create fft angular mask}
\usage{
fft_angular_mask(
  height,
  width,
  orientation = 0,
  angle = 0,
  cutoff = c(0, min(width, height)/2)
)
}
\arguments{
\item{height}{image height}

\item{width}{image width}

\item{orientation}{mask orientation in radians}

\item{angle}{angle of mask in radians}

\item{cutoff}{cutoff frequencies}
}
\value{
an image
}
\description{
Create fft angular mask
}
