% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fft_noise}
\alias{fft_noise}
\title{Generate a white noise image}
\usage{
fft_noise(height, width)
}
\arguments{
\item{height}{image height}

\item{width}{image width}
}
\value{
an image
}
\description{
Generate a white noise image
}
\examples{
plot(fft_noise(128, 128))
}
