% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fft_phase}
\alias{fft_phase}
\title{Calculate fft phase}
\usage{
fft_phase(im)
}
\arguments{
\item{im}{an image}
}
\value{
an image
}
\description{
Calculate fft phase
}
\examples{
phase = fft_phase(im_gray(regatta))
plot(phase, rescale = TRUE)
}
