% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fft_shift}
\alias{fft_shift}
\title{Shift fft image}
\usage{
fft_shift(im, inverse = FALSE)
}
\arguments{
\item{im}{an image}

\item{inverse}{inverse transform}
}
\value{
an image
}
\description{
Shift fft image
}
