% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fft_slope}
\alias{fft_slope}
\title{Calculate fourier slope}
\usage{
fft_slope(
  im,
  range = c(10, 512),
  n = 10,
  mask = TRUE,
  showplot = FALSE,
  rawdata = FALSE
)
}
\arguments{
\item{im}{an image}

\item{range}{frequency range}

\item{n}{number of samples}

\item{mask}{if TRUE, apply circular mask before calculation}

\item{showplot}{if TRUE, dislays a plot}

\item{rawdata}{if TRUE, raw data is returned as well as slope value}
}
\value{
a numeric value or a data frame
}
\description{
Calculate fourier slope
}
\examples{
fft_slope(regatta)
}
