% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fft_spectrum}
\alias{fft_spectrum}
\title{Calculate fft amplitude and phase}
\usage{
fft_spectrum(im, modify = FALSE)
}
\arguments{
\item{im}{an image}

\item{modify}{if TRUE, amplitude value is log-rescaled, which is useful for plotting purpose}
}
\value{
an image
}
\description{
Calculate fft amplitude and phase
}
\examples{
img = fft_spectrum(im_gray(regatta), modify = TRUE)
plot(img$magnitude)
plot(img$phase, rescale = TRUE)
}
