% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{filter_v1}
\alias{filter_v1}
\title{Apply V1 filters to an image}
\usage{
filter_v1(im, id, n_orientation = 4, cell_type = "complex")
}
\arguments{
\item{im}{an image}

\item{id}{a vector of integer. which cell to use.}

\item{n_orientation}{number of orientation}

\item{cell_type}{either "simple" or "complex" (default) cell}
}
\value{
a list of images
}
\description{
Apply V1 filters to an image
}
\examples{
im = filter_v1(im_gray(regatta), id = 1:2, n_orientation = 4, cell_type = "complex")
plot(im[[1]], rescale = TRUE)
}
