% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{Fractal2D}
\alias{Fractal2D}
\title{Calculate the fractal dimension of an image using the traditional 2D box counting method}
\usage{
Fractal2D(im, n = 20, thr = "auto", rawdata = FALSE)
}
\arguments{
\item{im}{an image}

\item{n}{number of samples}

\item{thr}{threshold for binarization. either numeric (e.g., 0.5), or "auto", or a string for quantiles (e.g., "15\%")}

\item{rawdata}{if TRUE, raw data is returned as well as slope value}
}
\value{
a numeric
}
\description{
Calculate the fractal dimension of an image using the traditional 2D box counting method
}
\examples{
Fractal2D(regatta)
}
