% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{gauss_kernel}
\alias{gauss_kernel}
\title{Create a Gaussian kernel}
\usage{
gauss_kernel(sd, radius = round(2.5 * sd))
}
\arguments{
\item{sd}{sd of the normal distribution}

\item{radius}{kernel radius. kernel diameter is 2 * radius + 1.}
}
\value{
an image
}
\description{
Create a Gaussian kernel
}
\examples{
plot(gauss_kernel(10), rescale = TRUE)
plot(gauss_kernel(sd = 10, radius = 20), rescale = TRUE)
}
