% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{get_alpha}
\alias{get_alpha}
\title{Extract the alpha channel from an image}
\usage{
get_alpha(im)
}
\arguments{
\item{im}{an image}
}
\value{
an image
}
\description{
Extract the alpha channel from an image
}
