% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{get_arc_mask}
\alias{get_arc_mask}
\title{Create an arc mask image}
\usage{
get_arc_mask(height, width, cy, cx, r, t1, t2)
}
\arguments{
\item{height}{image height}

\item{width}{image width}

\item{cy}{arc center y}

\item{cx}{arc center x}

\item{r}{arc radius}

\item{t1}{angle (start) in radius}

\item{t2}{angle (end) in radius}
}
\value{
an image
}
\description{
Create an arc mask image
}
\examples{
mask = get_arc_mask( 100, 120, 50, 50, 20, pi/6, pi )
plot(mask)
}
