% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{get_channel}
\alias{get_channel}
\title{Extract a color channel from an image}
\usage{
get_channel(im, channel)
}
\arguments{
\item{im}{an image}

\item{channel}{color channel to extract. either numeric or string.}
}
\value{
an image
}
\description{
Extract a color channel from an image
}
\examples{
R = get_channel(regatta, 1)
plot(R)
G = get_channel(regatta, "G")
plot(G)
}
