% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{get_circular_mask}
\alias{get_circular_mask}
\title{Create a circular mask image}
\usage{
get_circular_mask(height, width, cy, cx, r)
}
\arguments{
\item{height}{image height}

\item{width}{image width}

\item{cy}{circle center y}

\item{cx}{circle center x}

\item{r}{circle radius}
}
\value{
an image
}
\description{
Create a circular mask image
}
\examples{
mask = get_circular_mask( 100, 120, 50, 50, 20 )
plot(mask)
}
