% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{get_gradient}
\alias{get_gradient}
\title{Get a gradient image}
\usage{
get_gradient(im, starts.horizontal = TRUE)
}
\arguments{
\item{im}{an image}

\item{starts.horizontal}{axis direction}
}
\value{
a list of images
#' @examples
G = get_gradient(regatta)
plot(G$theta)
plot(G$magnitude)
}
\description{
Get a gradient image
}
