% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{get_hog}
\alias{get_hog}
\title{Calculate PHOG descriptor}
\usage{
get_hog(gradient, n.bin = 8, is.orientation = TRUE, starts.horizontal = TRUE)
}
\arguments{
\item{gradient}{gradient data}

\item{n.bin}{number of orientation/direction bins}

\item{is.orientation}{TRUE for orientation (0-180 degrees), FALSE for direction (0-360 degrees)}

\item{starts.horizontal}{if TRUE, the first bin is centered at 0 degree}
}
\value{
a data frame
}
\description{
Calculate PHOG descriptor
}
