% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{get_L}
\alias{get_L}
\title{Get L channel of CIELAB color space}
\usage{
get_L(im, scale = TRUE)
}
\arguments{
\item{im}{an image}

\item{scale}{if TRUE (default), L value is divided by 100}
}
\value{
an image
}
\description{
Get L channel of CIELAB color space
}
\examples{
plot(get_L(regatta))
}
