% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{get_moments}
\alias{get_moments}
\title{Get moment statistics of a vector/array}
\usage{
get_moments(x, order = 1:4, na.rm = FALSE)
}
\arguments{
\item{x}{data}

\item{order}{order of the moment to be computed}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
a numeric vector
}
\description{
Get moment statistics of a vector/array
}
\examples{
get_moments( rnorm( 20 ) ) # get the 1st to 4th order moments
get_moments( rnorm( 20 ), order = 3 ) # get only the 3rd order moment (skewness)
get_moments( rnorm( 20 ), order = c( 3, 1 ) ) # get skewness (3rd moment) and mean (1st moment)
}
