% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{get_oval_mask}
\alias{get_oval_mask}
\title{Create a oval mask image}
\usage{
get_oval_mask(height, width, cy, cx, ry, rx)
}
\arguments{
\item{height}{image height}

\item{width}{image width}

\item{cy}{oval center y}

\item{cx}{oval center x}

\item{ry}{oval radius y}

\item{rx}{oval radius x}
}
\value{
an image
}
\description{
Create a oval mask image
}
\examples{
mask = get_oval_mask(100, 200, 50, 10, 20, 40 )
plot(mask)
}
