% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_blend}
\alias{im_blend}
\title{Alpha blending of two images}
\usage{
im_blend(im, im2, alpha)
}
\arguments{
\item{im}{an image (foreground)}

\item{im2}{an image (background)}

\item{alpha}{an image (alpha channel). if missing, the alpha channel of im is used.}
}
\value{
an image
}
\description{
Alpha blending of two images
}
