% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_blur}
\alias{im_blur}
\title{Blur an image}
\usage{
im_blur(im, cutoff, method = "gauss", order = 4)
}
\arguments{
\item{im}{an image}

\item{cutoff}{cutoff frequency}

\item{method}{either "gauss", "ideal", or "butterworth"}

\item{order}{order of butterworth filter}
}
\description{
Blur an image
}
\examples{
plot(im_blur(regatta, 20))
}
