% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_combine}
\alias{im_combine}
\title{Combine images}
\usage{
im_combine(im1, im2, y = 0, x = 0, alpha = FALSE, background = 1)
}
\arguments{
\item{im1}{an image}

\item{im2}{an image}

\item{y}{y-offset}

\item{x}{x-offset}

\item{alpha}{either FALSE (default) or TRUE (enable alpha transparency)}

\item{background}{background color}
}
\value{
an image
}
\description{
Combine images
}
\examples{
plot(im_combine(regatta, regatta, y = im_height(regatta)))
plot(im_combine(regatta, regatta, y = 100, x = 200))
plot(im_combine(regatta, regatta, y = 100, x = 200, background = 0.5))
plot(im_combine(regatta, regatta, y = 100, x = 200, background = c(1, 0.5, 0.5)))
}
