% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_conv}
\alias{im_conv}
\title{Convolve an image}
\usage{
im_conv(im, kernel, pad.method = "mirror")
}
\arguments{
\item{im}{an image}

\item{kernel}{filter image}

\item{pad.method}{either "zero", "mean", "repeat", "mirror", or a numeric value}
}
\value{
an image
}
\description{
Convolve an image
}
\examples{
plot(im_conv(regatta, gauss_kernel(sd = 2)))
}
