% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_coord}
\alias{im_coord}
\title{Get pixel coordinates for an image, as an image}
\usage{
im_coord(im, direction = "x")
}
\arguments{
\item{im}{an image}

\item{direction}{Either "x" for x coordinates or "y" for y coordinates}
}
\value{
an image containing pixel coordinates
}
\description{
Get pixel coordinates for an image, as an image
}
\examples{
xc = im_coord(regatta, "x")
plot(xc, rescale = TRUE)
}
