% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_crop}
\alias{im_crop}
\title{Image cropping}
\usage{
im_crop(im, margin)
}
\arguments{
\item{im}{an image}

\item{margin}{a numeric vector.}
}
\value{
an image
}
\description{
Image cropping
}
\examples{
plot(im_crop(regatta, 100)) # crop all four sides with the same value
plot(im_crop(regatta, c(100, 50))) # vertical, horizontal
plot(im_crop(regatta, c(0, 50, 50, 200))) # top, right, bottom, left
}
