% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_crop_square}
\alias{im_crop_square}
\title{Make an image square shaped}
\usage{
im_crop_square(im, position = 0.5)
}
\arguments{
\item{im}{an image}

\item{position}{square center. a numeric value between 0 and 1.}
}
\value{
an image
}
\description{
Make an image square shaped
}
\examples{
plot(im_crop_square(regatta))
plot(im_crop_square(regatta, position = 0))
plot(im_crop_square(regatta, position = 0.8))
}
