% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_cy}
\alias{im_cy}
\title{Y coordinate of image center}
\usage{
im_cy(im)
}
\arguments{
\item{im}{an image}
}
\value{
Y coordinate of image center
}
\description{
When the height of the image is an even number, digits after a decimal point are rounded up.
}
\examples{
im_cy(regatta)
}
