% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_diagonal}
\alias{im_diagonal}
\title{Get the diagonal length of an image}
\usage{
im_diagonal(im)
}
\arguments{
\item{im}{an image}
}
\value{
image's diagonal length
}
\description{
Get the diagonal length of an image
}
\examples{
im_diagonal(regatta)
}
