% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_diff}
\alias{im_diff}
\title{Mean squared error between two images}
\usage{
im_diff(im1, im2)
}
\arguments{
\item{im1}{an image}

\item{im2}{an image}
}
\value{
mean squared error
}
\description{
Mean squared error between two images
}
\examples{
im_diff(regatta, regatta^2)
}
