% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_distribute}
\alias{im_distribute}
\title{Shift and scale the distribution of pixel values}
\usage{
im_distribute(
  im,
  channel,
  mean = NULL,
  sd = NULL,
  space = "CIELAB",
  clamp = TRUE
)
}
\arguments{
\item{im}{an image}

\item{channel}{color channel}

\item{mean}{center of distribution. when not given, the mean of that channel is used.}

\item{sd}{dispersion of distribution. when not given, the sd of that channel is used.}

\item{space}{color space}

\item{clamp}{either TRUE (default, output pixel value is clamped to range 0-1) or FALSE}
}
\value{
an image
}
\description{
Shift and scale the distribution of pixel values
}
\examples{
im_moments(regatta) # before manipulation
im_moments(im_distribute(regatta, "b", mean = 0, sd = 20)) # b channel is adjusted
plot(im_distribute(regatta, "b", mean = 0, sd = 20)) # see the effect
plot(im_distribute(regatta, c("a", "b"), c(-5, 0), c(15, 20))) # adjust two channels simultaneously
}
