% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_flip}
\alias{im_flip}
\title{Flip an image}
\usage{
im_flip(im, direction = "horizontal")
}
\arguments{
\item{im}{an image}

\item{direction}{Either "horizontal" (default) or "vertical".}
}
\value{
an image
}
\description{
Flip an image
}
\examples{
plot(im_flip(regatta, "horizontal"))
plot(im_flip(regatta, "vertical"))
}
