% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_flip_position}
\alias{im_flip_position}
\title{Flip an image}
\usage{
im_flip_position(im, direction = "horizontal", position)
}
\arguments{
\item{im}{an image}

\item{direction}{Either "horizontal" (default) or "vertical".}

\item{position}{where to flip the image}
}
\value{
an image
}
\description{
Flip an image
}
