% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_get}
\alias{im_get}
\title{Image cropping}
\usage{
im_get(im, y1, x1, y2, x2)
}
\arguments{
\item{im}{an image}

\item{y1}{y coordinate of top-left corner}

\item{x1}{x coordinate of top-left corner}

\item{y2}{y coordinate of bottom-right corner}

\item{x2}{x coordinate of bottom-right corner}
}
\value{
an image
}
\description{
Image cropping
}
\examples{
plot(im_get(regatta, 1, 1, 100, 100))
}
