% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_gray}
\alias{im_gray}
\title{Convert to grayscale}
\usage{
im_gray(im, tricolored = FALSE)
}
\arguments{
\item{im}{an image}

\item{tricolored}{if TRUE, returned image has three color channels}
}
\value{
an image
}
\description{
Convert to grayscale
}
\examples{
plot(im_gray(regatta))
}
