% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_mono}
\alias{im_mono}
\title{Create a monotone image}
\usage{
im_mono(color, height, width)
}
\arguments{
\item{color}{single numeric for grayscale images, or three numeric values for RGB (color) images.}

\item{height}{image height}

\item{width}{image width}
}
\value{
an image
}
\description{
Create a monotone image
}
\examples{
plot(im_mono(0.5, 128, 128)) # a gray square
plot(im_mono(c(1, 0.2, 0.2), 64, 128)) # a red rectangle
}
