% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_nc}
\alias{im_nc}
\title{The number of color channel of an image}
\usage{
im_nc(im)
}
\arguments{
\item{im}{an image}
}
\value{
the number of color channel
}
\description{
The number of color channel of an image
}
\examples{
im_nc(regatta)
}
