% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_noise}
\alias{im_noise}
\title{Create a white noise image}
\usage{
im_noise(height, width, n.channel = 3)
}
\arguments{
\item{height}{image height}

\item{width}{image width}

\item{n.channel}{number of color channel. either 1 (gras scale noise) or 3 (color noise)}
}
\value{
an image
}
\description{
Create a white noise image
}
\examples{
plot(im_noise(128, 128))
}
