% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_npix}
\alias{im_npix}
\title{Total number of pixels of an image}
\usage{
im_npix(im)
}
\arguments{
\item{im}{an image}
}
\value{
total number of pixels
}
\description{
Total number of pixels of an image
}
\examples{
im_npix(regatta)
}
