% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_pad}
\alias{im_pad}
\title{Pad image with n pixels}
\usage{
im_pad(im, n, method = "mirror")
}
\arguments{
\item{im}{an image}

\item{n}{number of pixels to pad with}

\item{method}{either "zero", "mean", "repeat", "mirror", or a numeric value}
}
\description{
Pad image with n pixels
}
\examples{
# zero padding
im = im_pad(regatta, 20, "zero")
plot(im)
# mirror padding
im = im_pad(regatta, 100, "mirror")
plot(im)
}
