% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_pile}
\alias{im_pile}
\title{Pile two images}
\usage{
im_pile(im, im2)
}
\arguments{
\item{im}{an image}

\item{im2}{an image}
}
\value{
an image
}
\description{
Pile two images
}
\examples{
plot(im_pile(regatta, im_gray(regatta)))
}
