% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_pixelate}
\alias{im_pixelate}
\title{Apply a mosaic effect}
\usage{
im_pixelate(im, size = 16)
}
\arguments{
\item{im}{an image}

\item{size}{box size in pixel (default is 16)}
}
\value{
an image
}
\description{
Apply a mosaic effect
}
\examples{
plot(im_pixelate(regatta, size = 16))
}
