% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_raise}
\alias{im_raise}
\title{Raise pixel values}
\usage{
im_raise(im, intercept)
}
\arguments{
\item{im}{an image}

\item{intercept}{a numeric value between 0 and 1}
}
\value{
an image
}
\description{
Raise pixel values
}
\examples{
plot(im_raise(regatta, 0.3))
}
