% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_random_pixelate}
\alias{im_random_pixelate}
\title{Apply a mosaic effect followed by position randomization}
\usage{
im_random_pixelate(im, size = 16, direction = "xy", noise = 1)
}
\arguments{
\item{im}{an image}

\item{size}{box size in pixel (default is 16)}

\item{direction}{shuffle direction (either "x", "y", or "xy"). default is "xy".}

\item{noise}{percent of noise (default is 1). a numeric value between 0 and 1.}
}
\value{
an image
}
\description{
Apply a mosaic effect followed by position randomization
}
\examples{
plot(im_random_pixelate(regatta, size = 16, direction = "xy", noise = 0.1))
}
