% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_randomize}
\alias{im_randomize}
\title{Apply a pixel randomization effect}
\usage{
im_randomize(
  im,
  size = 16,
  direction = "xy",
  shuffle = TRUE,
  fraction = 1,
  noise = 1
)
}
\arguments{
\item{im}{an image}

\item{size}{box size in pixel (default is 1)}

\item{direction}{shuffle direction (either "x", "y", or "xy"). default is "xy".}

\item{shuffle}{shuffle along the direction. either TRUE (default) or FALSE}

\item{fraction}{a fraction of stripes are shuffled. a numeric value between 0 and 1 (default is 1).}

\item{noise}{percent of noise. a numeric value between 0 and 1 (default is 1).}
}
\value{
an image
}
\description{
Apply a pixel randomization effect
}
\examples{
plot(im_randomize(regatta, size = 16, direction = "xy", shuffle = TRUE, fraction = 1, noise = 1))
}
