% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_rep}
\alias{im_rep}
\title{Replicate a channel along color dimension}
\usage{
im_rep(im, n = 3, channel = 1)
}
\arguments{
\item{im}{an image}

\item{n}{number of repeat}

\item{channel}{which channel to extract}
}
\value{
an image
}
\description{
Replicate a channel along color dimension
}
