% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_resize}
\alias{im_resize}
\title{Resize image}
\usage{
im_resize(im, height, width, interpolation = 1)
}
\arguments{
\item{im}{an image}

\item{height}{image height}

\item{width}{image width}

\item{interpolation}{Interpolation method. Either 0 (nearest-neighbor), 1 (linear), or 2 (cubic).}
}
\value{
an image
}
\description{
If either height or width is given, the other is determined to keep the aspect ratio of image.
}
\examples{
dim(regatta)
dim(im_resize(regatta, height = 150))
dim(im_resize(regatta, width = 300))
dim(im_resize(regatta, height = 100, width = 100))
}
