% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_resize_limit_min}
\alias{im_resize_limit_min}
\title{Resize image}
\usage{
im_resize_limit_min(im, bound, interpolation = 1)
}
\arguments{
\item{im}{an image}

\item{bound}{max image size (width/height)}

\item{interpolation}{Interpolation method. Either 0 (nearest-neighbor), 1 (linear), or 2 (cubic).}
}
\value{
an image
}
\description{
Resize image
}
\examples{
dim(regatta)
dim(im_resize_limit_min(regatta, 200))
}
