% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_resize_scale}
\alias{im_resize_scale}
\title{Resize image by a scale factor}
\usage{
im_resize_scale(im, scale = 1, interpolation = 1)
}
\arguments{
\item{im}{an image}

\item{scale}{a scale factor}

\item{interpolation}{Interpolation method. Either 0 (nearest-neighbor), 1 (linear), or 2 (cubic).}
}
\value{
an image
}
\description{
Resize image by a scale factor
}
\examples{
dim(regatta)
dim(im_resize_scale(regatta, 0.5)) # half size
}
