% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_reverse}
\alias{im_reverse}
\title{Reverse pixel values}
\usage{
im_reverse(im)
}
\arguments{
\item{im}{an image}
}
\value{
an image
}
\description{
Reverse pixel values
}
\examples{
plot(im_reverse(regatta))
}
