% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_rotate}
\alias{im_rotate}
\title{Rotate an image}
\usage{
im_rotate(
  im,
  angle,
  expand = FALSE,
  cx = NULL,
  cy = NULL,
  interpolation = 2,
  pad = "zero"
)
}
\arguments{
\item{im}{an image}

\item{angle}{rotation angle in degrees}

\item{expand}{either FALSE (default, does not change image size) or TRUE}

\item{cx}{center of rotation along x}

\item{cy}{center of rotation along y}

\item{interpolation}{Type of interpolation. Either 0 (nearest), 1 (linear), or 2 (cubic).}

\item{pad}{Type of padding. Either "zero", "neumann", or "repeat".}
}
\value{
an image
}
\description{
Rotate an image
}
\examples{
plot(im_rotate(regatta, 30))
plot(im_rotate(regatta, 30, expand = TRUE))
plot(im_rotate(regatta, 30, expand = TRUE, pad = "repeat"))
}
