% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_serialize}
\alias{im_serialize}
\title{Convert an image to a numeric vector
Pixel values of R channel comes first, and then G, and finally B.
Each channel is serialized vertically (column-wise).}
\usage{
im_serialize(im)
}
\arguments{
\item{im}{an image}
}
\description{
Convert an image to a numeric vector
Pixel values of R channel comes first, and then G, and finally B.
Each channel is serialized vertically (column-wise).
}
\examples{
x = im_serialize(regatta)
}
