% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_serialize_inv}
\alias{im_serialize_inv}
\title{Convert a serialized image to an image}
\usage{
im_serialize_inv(im, height, width, nc = 3)
}
\arguments{
\item{im}{serialized image array}

\item{height}{image height}

\item{width}{image width}

\item{nc}{number of channels}
}
\description{
Convert a serialized image to an image
}
\examples{
se = im_serialize(regatta)
im = im_serialize_inv(se, 450, 600)
sum((regatta - im)^2)
}
