% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_shift}
\alias{im_shift}
\title{Image shift operation}
\usage{
im_shift(im, axis = "x", lag = 0)
}
\arguments{
\item{im}{an image}

\item{axis}{either "x" or "y"}

\item{lag}{a numeric}
}
\value{
an image
}
\description{
Image shift operation
}
\examples{
# shift a image 100 pixels to the right
im = im_shift(regatta, axis = "x", lag = 100)
plot(im)
}
