% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_size}
\alias{im_size}
\title{Get image width and height}
\usage{
im_size(im)
}
\arguments{
\item{im}{an image}
}
\value{
image's height
}
\description{
Get image width and height
}
\examples{
im_size(regatta)
}
