% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_symmetry}
\alias{im_symmetry}
\title{Analyze image symmetry}
\usage{
im_symmetry(
  im,
  method = "both",
  direction = "both",
  sample.rate = 0.25,
  lowpass.cutoff,
  rawdata = FALSE
)
}
\arguments{
\item{im}{an image}

\item{method}{either "luminance", "edge", or "both"}

\item{direction}{either "horizontal", "vertical", or "both"}

\item{sample.rate}{a numeric between 0 and 1}

\item{lowpass.cutoff}{cutoff frequency for lowpass filtering}

\item{rawdata}{if FALSE, summary data is returned. if TRUE, rawdata is also returned}
}
\description{
Analyze image symmetry
}
\examples{
im = im_resize( regatta, 128 )
df = im_symmetry(im)
}
