% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_threshold}
\alias{im_threshold}
\title{Threshold grayscale image}
\usage{
im_threshold(im, thr = "auto", approx = TRUE, adjust = 1)
}
\arguments{
\item{im}{an image}

\item{thr}{a threshold}

\item{approx}{skip}

\item{adjust}{adjust the automatic threshold}
}
\value{
an image
}
\description{
Threshold grayscale image
}
\examples{
plot(im_threshold(get_R(regatta), thr = 0.6))
}
