% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{im_tricolored}
\alias{im_tricolored}
\title{Force an image to have three and only three color channels}
\usage{
im_tricolored(im)
}
\arguments{
\item{im}{an image}
}
\value{
an image
}
\description{
Force an image to have three and only three color channels
}
\examples{
R = get_R(regatta)
im = im_tricolored(R)
}
