% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{is.nimg}
\alias{is.nimg}
\title{Check if an object is a nimg object}
\usage{
is.nimg(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical. if an object is a nimg object.
}
\description{
Check if an object is a nimg object
}
\examples{
is.nimg(regatta)
}
