% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{Lab2sRGB}
\alias{Lab2sRGB}
\title{Lab to sRGB conversion}
\usage{
Lab2sRGB(im, use.D65 = TRUE)
}
\arguments{
\item{im}{an image}

\item{use.D65}{reference white, either TRUE (D65 is used) or FALSE (D50 is used).}
}
\value{
an image
}
\description{
Lab to sRGB conversion
}
