% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{logspace}
\alias{logspace}
\title{Generates a logarithmically spaced sequence}
\usage{
logspace(from, to, length.out)
}
\arguments{
\item{from}{starting value}

\item{to}{ending value}

\item{length.out}{number of intervening values}
}
\value{
a numeric vector
}
\description{
Generates a logarithmically spaced sequence
}
\examples{
logspace(10,100,10)
}
