% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/m4000.r
\name{m4000}
\alias{m4000}
\title{Pacific Islands, Countries, and Territories bathymetric data}
\format{A data frame:
\describe{
  \item{long}{points describing the longitude line segments}
  \item{lat}{points describing the latitude line segments}
}}
\source{
Bathymetry and coastline contours provided with this package are generated from SRTM30_Plus v8.0 global gridded data using the Generic Mapping Tools software package. See \url{ftp://topex.ucsd.edu/pub/srtm30_plus/}
}
\description{
A data set contain the 50 metre depth contour
}

