% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{merge_color}
\alias{merge_color}
\title{Merge a list of images into an image}
\usage{
merge_color(imlist)
}
\arguments{
\item{imlist}{a list of images}
}
\value{
an image
}
\description{
Merge a list of images into an image
}
\examples{
merge_color(split_color(regatta))
}
