% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{MinMax}
\alias{MinMax}
\title{Calculate the range of values}
\usage{
MinMax(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the range of x
}
\description{
Calculate the range of values
}
\examples{
MinMax(1:10)
}
