% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{nimg}
\alias{nimg}
\title{Create a nimg object}
\usage{
nimg(im, name)
}
\arguments{
\item{im}{a 3 dimensional numeric array}

\item{name}{a string for name attribute.}
}
\value{
an object of class nimg
}
\description{
nimg is a class for storing image data as a 3 dimensional array (y, x, c format).
y and x are spatial coordinates, and c is a color dimension.
}
\examples{
nimg(array(1,c(128,256,3)))
}
