% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{paint_arc}
\alias{paint_arc}
\title{Draw an arc to an image}
\usage{
paint_arc(im, cy, cx, r, w, t1, t2, color)
}
\arguments{
\item{im}{an image}

\item{cy}{arc center x}

\item{cx}{arc center y}

\item{r}{arc radius}

\item{w}{arc width}

\item{t1}{angle (start) in radius}

\item{t2}{angle (end) in radius}

\item{color}{circle color}
}
\value{
an image
}
\description{
Draw an arc to an image
}
\examples{
im = paint_arc(regatta, 200, 300, 100, 10, 0, pi, c(1, 0.2, 0.2))
plot(im)
}
