% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{paint_circle}
\alias{paint_circle}
\title{Draw a circle to an image}
\usage{
paint_circle(im, cy, cx, r, color)
}
\arguments{
\item{im}{an image}

\item{cy}{circle center x}

\item{cx}{circle center y}

\item{r}{circle radius}

\item{color}{circle color}
}
\value{
an image
}
\description{
Draw a circle to an image
}
\examples{
im = paint_circle(regatta, 100, 200, 20, c(1, 0.2, 0.2))
plot(im)
}
