% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{palette_visualize}
\alias{palette_visualize}
\title{Visualize the color palette of an image}
\usage{
palette_visualize(
  im,
  palette,
  type = "bar",
  threshold = 1,
  monospace = FALSE,
  width,
  size
)
}
\arguments{
\item{im}{an image}

\item{palette}{color palette}

\item{type}{either "bar" or "circle}

\item{threshold}{minor colors below this threshold is not displayed}

\item{monospace}{either TRUE or FALSE. If FALSE, If FALSE, the length of the color chart is
proportional to the percentage of occurrences of that color in the image.}

\item{width}{width of color chart. A numeric value between 0 and 1.}

\item{size}{output image resolution. Only works for the circle type. The default is 1024.}
}
\description{
Visualize the color palette of an image
}
\examples{
palette = palette_extract(regatta)
im = palette_visualize(regatta, palette, "bar")
plot(im)
im2 = palette_visualize(regatta, palette, "circle")
plot(im2)
}
