% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{PHOG_descriptor}
\alias{PHOG_descriptor}
\title{Calculate PHOG descriptor}
\usage{
PHOG_descriptor(
  gradient,
  level = 3,
  n.bin = 8,
  is.orientation = TRUE,
  starts.horizontal = TRUE,
  clockwise = FALSE
)
}
\arguments{
\item{gradient}{gradient data or input image}

\item{level}{number of PHOG pyramid levels}

\item{n.bin}{number of orientation/direction bins}

\item{is.orientation}{TRUE for orientation (0-180 degrees), FALSE for direction (0-360 degrees)}

\item{starts.horizontal}{if TRUE, the first bin is centered at 0 degree}

\item{clockwise}{direction of rotation}
}
\value{
a data frame
}
\description{
Calculate PHOG descriptor
}
\examples{
df = PHOG_descriptor(regatta, level = 2)
}
