% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{PHOG_features}
\alias{PHOG_features}
\title{Calculate PHOG features}
\usage{
PHOG_features(
  im,
  level = 3,
  n.bin = 8,
  is.orientation = TRUE,
  starts.horizontal = TRUE,
  level.weight = c(1, 1, 1)
)
}
\arguments{
\item{im}{an image}

\item{level}{number of PHOG pyramid levels}

\item{n.bin}{number of orientation/direction bins}

\item{is.orientation}{TRUE for orientation (0-180 degrees), FALSE for direction (0-360 degrees)}

\item{starts.horizontal}{if TRUE, the first bin is centered at 0 degree}

\item{level.weight}{relative weights to each level for calculating self-similarity}
}
\value{
a data frame
}
\description{
Calculate PHOG features
}
\examples{
PHOG_features(regatta)
}
