% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{pow}
\alias{pow}
\title{Power operation}
\usage{
pow(x, p)
}
\arguments{
\item{x}{a numeric vector}

\item{p}{power term}
}
\value{
a numeric vector
}
\description{
Power operation
}
\examples{
pow(2, 3)
}
