% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.addaxis}
\alias{pp.addaxis}
\title{Add an axis to a PICT plot}
\usage{
pp.addaxis(side = 1, labels.type, cex, lty, mgp, tcl, minor.tcl,
  label.text = TRUE, hadj, padj, inside, ...)
}
\arguments{
\item{side}{side to add the axis}

\item{labels.type}{}

\item{cex}{cex}

\item{lty}{lty}

\item{mgp}{mgp}

\item{tcl}{tcl}

\item{label.text}{label.text}

\item{labels.type}{labels.type}

\item{minor.tcp}{minor.tcp}
}
\value{
Invisibly return axis locations of major ticks
}
\description{
Adds an axis to a plot generated by PICT, allowing full freedom to format an axis and tick mark labels.
}
\details{
Adds an axis to a pp() plot
}

