% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.arrows}
\alias{pp.arrows}
\title{Add arrows to a PICT plot}
\usage{
pp.arrows(x1, y1, x2, y2, length = 0.1, ...)
}
\arguments{
\item{x1,}{y1 coordinates of points to which to draw. At least one must the supplied}

\item{length}{length of the edges of the arrow head (in inches).}

\item{...}{arguments passes to arrows()}

\item{x0,}{y0 coordinates of points from which to draw.}
}
\value{
Returns no value
}
\description{
Adds arrows to a PICT plot
}
\details{
Adds arrows to a pp() plot
}

