% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.bubble}
\alias{pp.bubble}
\title{Add a bubble plot to a PICT plot}
\usage{
pp.bubble(x, y = NULL, radius, scale = NULL, plot.centre = FALSE,
  resolution = 360, border = par("fg"), col = NA, density = NULL,
  angle = 45, ...)
}
\description{
Add a bubble plot to a PICT plot
}

