% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.circle}
\alias{pp.circle}
\title{Add constant distance circle to a PICT plot}
\usage{
pp.circle(distance, long, lat, plot = TRUE, fill.col = 0, fill = !fill.col
  \%in\% 0, ...)
}
\arguments{
\item{distance}{The circle radius in nautical miles (multiply by 1.852 to get kilometers)}

\item{long,}{lat Location of the circle centre in decimal degrees}

\item{plot}{if plot = FALSE, the function will return the point co-ordinates, and not plot}

\item{fillcol}{if fillcol != 0, then then circle will be filled with colour fill.col}
}
\value{
Returns no value
}
\description{
Draws circles with specified great circle radii (in nautical miles) around specified centres
}

