% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pp.image.legend.R
\name{pp.image.legend}
\alias{pp.image.legend}
\title{Add an image legend to a PICT plot}
\usage{
pp.image.legend(zmat, x, y, size = c(0.25, 2), horizontal = F, nint = 50,
  plus = F, lab, ...)
}
\arguments{
\item{...}{arguments passed to image()}
}
\value{
Returns no value
}
\description{
See R graphics function image for details
}

