% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pp.legend.R
\name{pp.legend}
\alias{pp.legend}
\title{Add a legend to a PICT plot}
\usage{
pp.legend(x, y, ...)
}
\arguments{
\item{x,}{y  the x and y co-ordinates for the location of legend}

\item{...}{arguments passes to legend()}
}
\value{
Returns no value
}
\description{
Adds a legend to a plot.
}

