% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pp.lines.R
\name{pp.lines}
\alias{pp.lines}
\title{Add lines to a PICT plot}
\usage{
pp.lines(x, y = NULL, ...)
}
\arguments{
\item{x,}{y  numeric vectors of coordinates where the lines should be written. If the length of x and y differs, the shorter one is recycled.}
}
\value{
Returns no value
}
\description{
Draws lines whose vertices are given in x and y.
}
\details{
Adds a line to a pp() plot
}

