% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pp.locator.R
\name{pp.locator}
\alias{pp.locator}
\title{return a lat/long location from a PICT plot}
\usage{
pp.locator(...)
}
\arguments{
\item{...}{additional parameters passes to locator()}
}
\value{
Returns a list of lat/long points identifed by the cursor
}
\description{
Uses the cursor to identify a lat/long location on a PICT plot
}

