% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.north}
\alias{pp.north}
\title{Add north arrow to a PICT plot}
\usage{
pp.north(x, y, size = 0.1, label = T, length = 0.25, angle = 30,
  code = 2, ...)
}
\arguments{
\item{x,}{y  numeric vectors of coordinates where tarrow should be written}

\item{size}{size (in proportion of y limits of the plot) or the arrow}

\item{label}{if true, then adds a label with 'N' at the top of the arrow}

\item{length}{}

\item{x,}{y}
}
\description{
Draws an arrow indicating north
}

