% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.pie}
\alias{pp.pie}
\title{Add a pie chart to a PICT plot}
\usage{
pp.pie(x, y = NULL, z, size = 0.2, resolution = 360, start.angle = 1, col = rainbow(length(z)), border = T, ...)
}
\arguments{
\item{x,}{y  numeric vectors of coordinates where the text labels should be written. If the length of x and y differs, the shorter one is recycled.}

\item{z}{a vector of relative sizes of each slice (unscaled)}

\item{size}{diameter of the pie as a proportion of the length of the x axis}

\item{resolution}{an integer specifying the numbers of points used to draw the circle. Default is 360, which corresponds to a point every degree around the circle.}

\item{x,}{y x, y give the location of the centre of the pie charts. x is either a scalar giving the longitude of the centre or a vector of length 2 giving the long and lat (in that order) of the centre}
}
\value{
Returns no value
}
\description{
Draws pie charts at specified locations on a map drawn by PICTplot using either the Mercator or rectangular projection.
}
\details{
Adds a pie graph onto a pp() plot
}

