% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pp.points.R
\name{pp.points}
\alias{pp.points}
\title{Add points to a PICT plot}
\usage{
pp.points(x, y = NULL, ...)
}
\arguments{
\item{x,}{y  the x and y co-ordinates for the centres of the symbols.}
}
\value{
Returns no value
}
\description{
TDraws draw a sequence of points at the specified coordinates.
}
\details{
Adds points to a pp() plot
}

