% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.polygon}
\alias{pp.polygon}
\title{Add a polygon to a PICT plot}
\usage{
pp.polygon(x, y = NULL, ...)
}
\arguments{
\item{x,}{y  the x and y co-ordinates for the centres of the symbols.}
}
\value{
Returns no value
}
\description{
Draws polygons whose vertices are given in x and y.
}
\details{
Adds a polygon to a pp() plot
}

