% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp}
\alias{pp}
\title{PICTplot high level plotting function}
\usage{
pp(xlim = c(140, 225), ylim = c(10, -30), coastline = TRUE, lty = 1,
  lwd = 1, col = 1, fill.col = 0, density = NA, angle = 45,
  projection = "merc", aspect = "ratio", axes = TRUE, xaxt, yaxt,
  tcl = -0.3, minor.tcl = 0, labels = "degree", cex = 1, mgp, las,
  xlab = NULL, ylab = NULL, main = NULL, sub = NULL, box.lty = 1,
  box.lwd = 1, box.col = 1, background.col, border = 0, grid = F, ...)
}
\arguments{
\item{xlim}{vector containing the plot x limits in decimal degrees with longitudes in the range 0 to 360.}

\item{ylim}{vector containing the plot y limits in decimal degrees with longitudes in the range 0 to 360.}

\item{coastline}{logicals determining whether the island coastline are to be drawn and/or filled.}

\item{lty}{line type of the borders of the land features drawn.}

\item{lwd}{line width of the borders of the land features drawn.}

\item{col}{line colour of the borders of the land features drawn.}

\item{fill.col}{polygon drawing arguments. Use fill.col = 0 and density = NA to ensure that no polygon is drawn. If fill.col is not 0 then polygon is filled with fill.col and/or if density is numeric then angled lines (density gives number of lines per inch and angle gives their angle. See the R Graphics function polygon for parameter settings and their effects).}

\item{projection}{character string, either "merc" or "rect", (mercator or rectangular).}

\item{aspect}{character string, one of "ratio", "square", or "maximum". "ratio" preserves the ratio of the lengths of the x and y axes. "square" always plots in a square. "maximum" uses all the available plot area without fixing the NS to EW aspect ratio. Default is aspect = "ratio".}

\item{axes}{arguments determining whether both or neither axes are drawn.}

\item{xaxt}{If axes is FALSE then neither axis is drawn and no box is placed around the plot area. If axes = TRUE the axes are drawn on the bottom and left of the figure unless overruled by xaxt = "n" and/or yaxt = "n" (which enables only one axis to be drawn).}

\item{yaxt}{If axes is FALSE then neither axis is drawn and no box is placed around the plot area. If axes = TRUE the axes are drawn on the bottom and left of the figure unless overruled by xaxt = "n" and/or yaxt = "n" (which enables only one axis to be drawn).}

\item{tcl}{determine the length of the major tick marks (typically whole degrees) and minor.tcl the length of the minor ticks marks (associated with numbers of minutes) and in the fraction of a mex unit. Negative values put ticks outside the axis and positive inside. If minor.tcl = 0 the minor ticks are not drawn. If tcl = 0, then no tick marks are drawn (no matter what minor.tcl is set to).}

\item{labels}{logical or a character string governing the printing of tick labels: one of "degree" , "decimal", "inside", TRUE, or FALSE. If "degree" or "decimal" then the tick labels are written in the named format outside the plot area. If "inside" tick labels are written inside the plot area. If FALSE (or "FALSE") then no labels are written. If TRUE, then labels are outside in degree tick format. The position of tick labels in relation to the axis is overruled by the second component of the mgp argument.}

\item{cex}{character expansion for tick labels.}

\item{mgp}{graphics parameter, a numeric vector of length 3, which sets, for each side, the number of lines from the plot edge into the margin that: the axis title is written, the tick labels are written, and the axis position is drawn (see par and axis functions help in the R Graphics package). If mgp is missing (default), PICT draws the axes and positions the labels according to internal defaults (axes are drawn on plot area border, and labels are placed on the line so there is a gap of 0.5*cex lines between the line and the ends of the major tick marks. Setting mgp will overrule the inside/outside setting for the labels argument depending on whether the second component of mgp is less than 0. Setting mgp also changes the line where the axes titles are drawn, (mgp[1]) and also the line position where the axes are drawn, (mgp[3]).}

\item{las}{graphics parameter that determines the orientation of the tick labels. Restricted in PICT to only 0 or 1. If not set, the default value is las = 1, labels are are horizontal for both axes.}

\item{xlab}{character arguments which give the axes titles, the main title,}

\item{main,}{sub and the sub-title.}

\item{box.lty}{line type, width, and colour for the box around the plot.}

\item{box.lwd}{box.lty = 0 supresses the drawing of the box}

\item{box.col}{Box colour}

\item{background.col}{Either missing or the background colour to shade the whole plot area. If missing, there is no shading (transparent).#'}

\item{border}{Extent of border around outside of plot. default=0}

\item{grid}{Add 5 degree grid lines (default is false). See also pp.grid()}

\item{...}{Additional plotting parameters passed to some plotting commands within .PICT (For further details see par().) Examples: cex.main, col.main, font.main for character expansion, colour, and font of the main title and cex.lab for character expansion for the xlab and ylab titles.}

\item{coastline}{logical determining whether coastline is plotted (or filled)}

\item{xlab}{character arguments which give the axes titles, the main title,}
}
\value{
Returns no value but the side effect is to draw the plot area.
}
\description{
High level map plotting function for the lands and oceans in the PICTn Pacific region.
The plotting parameters, set in pp, are saved in the list .PICT in the
global environment, for use as required with some low level plotting
functions.
}
\details{
After setting up plot area, then if required, PICT fills the plot area with background.col, fills land areas using pp.polygon and draws coastlines using pp.lines. It next draws the axes using pp.labels.degree of pp.labels.decimal and a box is drawn around the plot area. If required, a title, subtitle and axis titles are added and the size, colour, and font can be controlled by cex.main, cex.sub, cex.lab, col.main, etc. which are passed via the ... argument.  The characteristics of the plot are saved in the list object .PICT in the global environment for later use by low level plotting functions that add features to the exsting plot.
}
\author{
Alistair Dunn
}

