% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pp.segments.R
\name{pp.segments}
\alias{pp.segments}
\title{Add segments to a PICT plot}
\usage{
pp.segments(x0, y0, x1, y1, ...)
}
\arguments{
\item{x0,}{y0  coordinates of points from which to draw.}

\item{x1,}{y1  coordinates of points to which to draw. At least one must be supplied.}
}
\value{
Returns no value
}
\description{
Draws line segments between pairs of points.
}
\details{
Adds segments to a pp() plot
}

