% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pp.symbols.R
\name{pp.symbols}
\alias{pp.symbols}
\title{Add symbols to a PICT plot}
\usage{
pp.symbols(x, y = NULL, add = TRUE, ...)
}
\arguments{
\item{x,}{y  the x and y co-ordinates for the centres of the symbols.}
}
\value{
Returns no value
}
\description{
Draws symbols on a plot. One of six symbols; circles, squares, rectangles, stars, thermometers, and boxplots, can be plotted at a specified set of x and y coordinates. Specific aspects of the symbols, such as relative size, can be customized by additional parameters.
}
\details{
Adds symbols to a pp() plot
}

