% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pp.text.R
\name{pp.text}
\alias{pp.text}
\title{Add text to a PICT plot}
\usage{
pp.text(x, y = NULL, ...)
}
\arguments{
\item{x,}{y  numeric vectors of coordinates where the text labels should be written. If the length of x and y differs, the shorter one is recycled.}
}
\value{
Returns no value
}
\description{
Draws the strings given in the vector labels at the coordinates given by x and y. y may be missing since xy.coords(x, y) is used for construction of the coordinates.
}
\details{
Adds text to a pp() plot
}

