% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{pplot}
\alias{pplot}
\title{Display an image}
\usage{
pplot(im, rescale = FALSE)
}
\arguments{
\item{im}{an image. either nimg object or an array object}

\item{rescale}{logical. if true, then pixel value is rescaled to range between 0 and 1.}
}
\value{
No return value, called for side effects.
}
\description{
Display an image
}
\examples{
pplot(regatta)
}
