% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{ramp_threshold}
\alias{ramp_threshold}
\title{Apply a thresholding function}
\usage{
ramp_threshold(x, eta, phi)
}
\arguments{
\item{x}{a numeric vector}

\item{eta}{threshold}

\item{phi}{smoothing parameter}
}
\value{
a numeric vector
}
\description{
For details, see http://dx.doi.org/10.1016/j.cag.2012.03.004
}
\examples{
x = seq( from = 0, to = 1, by = 0.01 )
plot(x, ramp_threshold( x, 0.5, 6 ))
}
