% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regatta}
\alias{regatta}
\title{An image.}
\format{
An array with 450 x 600 x 3 dimensions.
Each dimension represents y-coordinate, x-coordinate, and color channel.
}
\usage{
regatta
}
\description{
A painting by Claude Monet.
}
\examples{
plot(regatta)
}
\keyword{datasets}
