% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{remove_alpha}
\alias{remove_alpha}
\title{Remove alpha channel from an image}
\usage{
remove_alpha(im, index)
}
\arguments{
\item{im}{an image}

\item{index}{alpha channel index. 4 for RGB image, 2 for grayscale image.}
}
\value{
an image
}
\description{
Remove alpha channel from an image
}
