% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{remove_channel}
\alias{remove_channel}
\title{Remove a chennel(s) from an image}
\usage{
remove_channel(im, index)
}
\arguments{
\item{im}{an image}

\item{index}{channel index to be removed}
}
\value{
an image
}
\description{
Remove a chennel(s) from an image
}
\examples{
plot(remove_channel(regatta, 1:2)) # remove 1st (R) and 2nd (G) channels
}
