% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{rescaling}
\alias{rescaling}
\title{Rescale numeric vector to have a specified range}
\usage{
rescaling(x, from = 0, to = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{from}{lowest value}

\item{to}{highest value}
}
\value{
a rescaled numeric vector
}
\description{
Rescale numeric vector to have a specified range
}
\examples{
rescaling( 1:5, from = 0, to = 10 )
}
\seealso{
\code{\link{rescaling01}}
}
