% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{rescaling01}
\alias{rescaling01}
\title{Rescale numeric vector to have a range between 0 to 1}
\usage{
rescaling01(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a rescaled numeric vector
}
\description{
Rescale numeric vector to have a range between 0 to 1
}
\examples{
rescaling01( 1:5 )
}
\seealso{
\code{\link{rescaling}}
}
