% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{RGB2Lab}
\alias{RGB2Lab}
\title{RGB to Lab conversion}
\usage{
RGB2Lab(im, use.D65 = TRUE)
}
\arguments{
\item{im}{an image}

\item{use.D65}{reference white, either TRUE (D65 is used) or FALSE (D50 is used).}
}
\value{
an image
}
\description{
RGB to Lab conversion
}
