% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{RGB2XYZ}
\alias{RGB2XYZ}
\title{RGB to XYZ conversion}
\usage{
RGB2XYZ(im, use.D65 = TRUE)
}
\arguments{
\item{im}{an image}

\item{use.D65}{reference white, either TRUE (D65 is used) or FALSE (D50 is used).}
}
\value{
an image
}
\description{
RGB to XYZ conversion
}
