% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{smart.round}
\alias{smart.round}
\title{Round floats to integers while preserving their sum
ref. https://stackoverflow.com/questions/32544646/round-vector-of-numerics-to-integer-while-preserving-their-sum}
\usage{
smart.round(x, digits = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places (round) to be used.}
}
\value{
a numeric vector
}
\description{
Round floats to integers while preserving their sum
ref. https://stackoverflow.com/questions/32544646/round-vector-of-numerics-to-integer-while-preserving-their-sum
}
\examples{
smart.round(seq(1,1.5,0.05), 1)
}
