% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{sRGB2Lab}
\alias{sRGB2Lab}
\title{sRGB to Lab conversion}
\usage{
sRGB2Lab(im, use.D65 = TRUE)
}
\arguments{
\item{im}{an image}

\item{use.D65}{reference white, either TRUE (D65 is used) or FALSE (D50 is used).}
}
\value{
an image
}
\description{
sRGB to Lab conversion
}
