% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{sRGB2YUV}
\alias{sRGB2YUV}
\title{sRGB to YUV conversion}
\source{
\url{ https://en.wikipedia.org/wiki/YUV }
}
\usage{
sRGB2YUV(im, use.B601 = FALSE)
}
\arguments{
\item{im}{an image}

\item{use.B601}{logical. Either TRUE (SDTV with BT.601) or FALSE (HDTV with BT.709).}
}
\value{
an image
}
\description{
sRGB to YUV conversion
}
