% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{stat_filter}
\alias{stat_filter}
\title{Apply statistical filter}
\usage{
stat_filter(im, radius, FUN, pad.method = "mirror")
}
\arguments{
\item{im}{an image}

\item{radius}{kernel radius}

\item{FUN}{e.g., min, max, median, mean, var}

\item{pad.method}{either "zero", "mean", "repeat", "mirror", or a numeric value}
}
\value{
an image
}
\description{
Apply statistical filter
}
\examples{
plot(stat_filter(regatta, 1, min))
}
