% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{v1_kernels}
\alias{v1_kernels}
\title{Create a Gabor filter bank}
\usage{
v1_kernels(id, n_orientation = 4, cell_type = "complex")
}
\arguments{
\item{id}{a vector of integer. which cell to use.}

\item{n_orientation}{number of orientation}

\item{cell_type}{either "simple" or "complex" (default) cell}
}
\value{
a list of filter kernels
}
\description{
Create a Gabor filter bank
}
