% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{vec_shift}
\alias{vec_shift}
\title{Shift operation}
\usage{
vec_shift(v, lag = 0)
}
\arguments{
\item{v}{a numeric vector}

\item{lag}{a numeric}
}
\value{
a lagged vector
}
\description{
Shift operation
}
\examples{
vec_shift(1:7, 2)
}
