% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{XYZ2RGB}
\alias{XYZ2RGB}
\title{XYZ to RGB conversion}
\usage{
XYZ2RGB(im, use.D65 = TRUE)
}
\arguments{
\item{im}{an image}

\item{use.D65}{reference white, either TRUE (D65 is used) or FALSE (D50 is used).}
}
\value{
an image
}
\description{
XYZ to RGB conversion
}
