% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{XYZ2sRGB}
\alias{XYZ2sRGB}
\title{XYZ to sRGB conversion}
\usage{
XYZ2sRGB(im, use.D65 = TRUE)
}
\arguments{
\item{im}{an image}

\item{use.D65}{reference white, either TRUE (D65 is used) or FALSE (D50 is used).}
}
\value{
an image
}
\description{
XYZ to sRGB conversion
}
