% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{DBC}
\alias{DBC}
\title{Calculate fractal dimension}
\usage{
DBC(im, method = "DBC", n = 20)
}
\arguments{
\item{im}{an image}

\item{method}{either "DBC", "RDBC", or "IDBC"}

\item{n}{number of samples}
}
\value{
a numeric
}
\description{
Calculate fractal dimension
}
\examples{
DBC(regatta)
}
