% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictplot.R
\name{fft_slope}
\alias{fft_slope}
\title{Calculate orientation-averaged amplitude}
\usage{
fft_slope(
  im,
  range = c(10, 512),
  n = 10,
  mask = TRUE,
  linear.sample = TRUE,
  rawdata = FALSE,
  show.plot = FALSE
)
}
\arguments{
\item{im}{an image}

\item{range}{frequency range}

\item{n}{number of samples}

\item{mask}{if TRUE, apply circular mask before calculation}

\item{linear.sample}{if TRUE,}

\item{rawdata}{if TRUE, raw data is returned as well as slope value}

\item{show.plot}{if TRUE, figure is displayed}
}
\value{
a data frame
}
\description{
Calculate orientation-averaged amplitude
}
\examples{
fft_slope(regatta)
fft_slope(regatta, rawdata = TRUE, show.plot = TRUE)
}
