% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.addaxis}
\alias{pp.addaxis}
\title{Add an additional axis to a PICT plot}
\usage{
pp.addaxis(side = 1, labels.type, cex, lty, mgp, tcl, minor.tcl,
  label.text = TRUE, hadj, padj, inside, ...)
}
\description{
Add an additional axis to a PICT plot
}

