% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.depth}
\alias{pp.depth}
\title{Add depth contours to a PICT plot}
\usage{
pp.depth(contour = NULL, lty = 1, lwd = 1, col = 1)
}
\arguments{
\item{contour}{The depth contours to be plotted}
}
\value{
Returns no value
}
\description{
Adds depth contours to a PICT plot
}
\details{
Adds a bathymetric contour to a pp() plot
}

