% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.grid}
\alias{pp.grid}
\title{Add axis grid lines to a PICT plot}
\usage{
pp.grid(at = 5, col = "gray", lty = 1, ...)
}
\arguments{
\item{at}{the frequency (in degrees) at which gridlines will be drawn. Defaults to 5 degrees}

\item{...}{other arguments to be passed to the R graphics package function lines}
}
\value{
Returns no value
}
\description{
Adds grid lines to a PICT plot
}
\details{
Grid lines are added to the plot at the specified interval
}

