% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.image.legend}
\alias{pp.image.legend}
\title{Add an image legend to a PICT plot}
\usage{
pp.image.legend(zmat, x, y, size = c(0.25, 2), horizontal = F, nint = 50,
  plus = F, lab, ...)
}
\description{
See R graphics function image for details of parameters that are passed to image by ...
}

