% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pp.legend}
\alias{pp.legend}
\title{Add a legend to a PICT plot}
\usage{
pp.legend(x, y, ...)
}
\arguments{
\item{x,}{y  the x and y co-ordinates for the location of legend}

\item{...}{arguments passes to legend()}
}
\value{
Returns no value
}
\description{
Adds a legend to a plot.
}

